Created a brand new DuckDB configuration to replace the old #738 pull
request.

This should now support the most important bits of DuckDB. There are
some caveats though:

- No support for the percentage syntax, like `LIMIT 10%` (conflicts with
modulo operator).
- No support for named parameters like `$foo` (conflicts with $$-quoted
strings).
- No support for array-slice operator (conflicts with `:` in struct
literals and prefix aliases).

There's definitely quite a bit more than the above three little things.
But I think at least for start we have some DuckDB support that should
work for most users.

+++ test/features/truncateTable.ts
@@ -3,19 +3,22 @@ import dedent from 'dedent-js';
 import { FormatFn } from '../../src/sqlFormatter.js';
 
 interface TruncateTableConfig {
+  withTable?: boolean;
   withoutTable?: boolean;
 }
 
 export default function supportsTruncateTable(
   format: FormatFn,
-  { withoutTable }: TruncateTableConfig = {}
+  { withTable = true, withoutTable }: TruncateTableConfig = {}
 ) {
-  it('formats TRUNCATE TABLE statement', () => {
-    const result = format('TRUNCATE TABLE Customers;');
-    expect(result).toBe(dedent`
-      TRUNCATE TABLE Customers;
-    `);
-  });
+  if (withTable) {
+    it('formats TRUNCATE TABLE statement', () => {
+      const result = format('TRUNCATE TABLE Customers;');
+      expect(result).toBe(dedent`
+        TRUNCATE TABLE Customers;
+      `);
+    });
+  }
 
   if (withoutTable) {
     it('formats TRUNCATE statement (without TABLE)', () => {
